gcf/restore-process/src/requirements.txt
pandas == 2.0.0
google-cloud-secret-manager == 2.16.1
google-cloud-storage == 2.8.0
fastavro == 1.7.3
psycopg2-binary == 2.9.6
SQLAlchemy == 2.0.10
functions-framework == 3.3.0
